MxShop/requirements.txt
django
pillow
mysqlclient
djangorestframework-jwt
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven
django-guardian
social-auth-app-django

django-jet
django-grappelli
google-api-python-client==1.4.1
feedparser
django-filebrowser

pyprof2calltree
html-testRunner

# 性能分析工具
django-debug-toolbar
pyinstrument
django-extensions
